pyproject.toml
[tool.poetry]
name = "closet"
version = "0.1.0"
description = ""
authors = ["Nealon Hager <nealonhager@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^11.3.0"
click = "^8.2.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.13.0"
google-genai = "^1.36.0"
icecream = "^2.1.7"
pre-commit = "^4.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
